droby: move code around to have a one-class-per-file scheme
