+++ lib/roby/distributed/notifications.rb
@@ -182,115 +182,6 @@ module Roby
 	end
 	Plan.include PlanModificationHooks
 
-	class PeerServer
-            # Message received when +task+ has become a mission (flag = true),
-            # or has become a non-mission (flag = false) in +plan+.
-	    def plan_set_mission(plan, task, flag)
-		plan = peer.local_object(plan)
-		task = peer.local_object(task)
-		if plan.owns?(task)
-		    if flag
-			plan.add_mission(task)
-		    else
-			plan.remove_mission(task)
-		    end
-		else
-		    task.mission = flag
-		end
-		nil
-	    end
-
-            # Message received when the set of tasks +m_tasks+ has been
-            # added by the remote plan. +m_relations+ describes the
-            # internal relations between elements of +m_tasks+. It is in a
-            # format suitable for PeerServer#set_relations.
-	    def plan_add(plan, m_tasks, m_relations)
-		Distributed.update(plan = peer.local_object(plan)) do
-		    tasks = peer.local_object(m_tasks).to_value_set
-		    Distributed.update_all(tasks) do 
-			plan.add(tasks)
-			m_relations.each_slice(2) do |obj, rel|
-			    set_relations(obj, rel)
-			end
-		    end
-		end
-		nil
-	    end
-
-            # Message received when +m_from+ has been replaced by +m_to+ in the
-            # plan
-	    def plan_replace(plan, m_from, m_to)
-		Distributed.update(plan = peer.local_object(plan)) do
-		    from, to = peer.local_object(m_from), peer.local_object(m_to)
-
-		    Distributed.update_all([from, to]) { plan.replace(from, to) }
-
-		    # Subscribe to the new task if the old task was subscribed
-		    # +from+ will be unsubscribed when it is finalized
-		    if peer.subscribed?(from) && !peer.subscribed?(to)
-			peer.subscribe(to)
-			nil
-		    end
-		end
-		nil
-	    end
-
-            # Message received when +object+ has been removed from +plan+
-	    def plan_remove_object(plan, object)
-		if local = peer.local_object(object, false)
-		    # Beware, transaction proxies have no 'plan' attribute
-		    plan = peer.local_object(plan)
-		    Distributed.update(plan) do
-			Distributed.update(local) do
-			    plan.remove_object(local)
-			end
-		    end
-		    local.forget_peer(peer)
-		end
-
-	    rescue ArgumentError => e
-		if e.message =~ /has not been included in this plan/
-		    Roby::Distributed.warn "filtering the 'not included in this plan bug'"
-		else
-		    raise
-		end
-	    end
-
-            # Message received when a relation graph has been updated. +op+ is
-            # either +add_child_object+ or +remove_child_object+ and describes
-            # what relation modification should be done. The two plan objects
-            # +m_from+ and +m_to+ are respectively linked or unlinked in the
-            # relation +m_rel+, with the given information object in case of a
-            # new relation.
-	    def update_relation(plan, m_from, op, m_to, m_rel, m_info = nil)
-		if plan
-		    Roby::Distributed.update(peer.local_object(plan)) { update_relation(nil, m_from, op, m_to, m_rel, m_info) }
-		else
-		    from, to = 
-			peer.local_object(m_from, false), 
-			peer.local_object(m_to, false)
-
-		    if !from
-			return unless to && (to.self_owned? || to.subscribed?)
-			from = peer.local_object(m_from)
-		    elsif !to
-			return unless from && (from.self_owned? || from.subscribed?)
-			to   = peer.local_object(m_to)
-		    end
-
-		    rel = peer.local_object(m_rel)
-		    Roby::Distributed.update_all([from.root_object, to.root_object]) do
-			if op == :add_child_object
-			    from.add_child_object(to, rel, peer.local_object(m_info))
-			elsif op == :remove_child_object
-			    from.remove_child_object(to, rel)
-			end
-		    end
-		end
-		nil
-	    end
-	end
-
         # This module defines the hooks needed to notify our peers of relation
         # modifications. It is included in plan objects.
 	module RelationModificationHooks
@@ -396,78 +287,6 @@ module Roby
 	end
 	Roby::EventGenerator.include EventNotifications
 
-	class PeerServer
-            # A set of events which have been received by #event_fired. This
-            # cache in cleaned up by PlanCacheCleanup#finalized_event when the
-            # associated generator is finalized.
-            #
-            # This cache is used to merge the events between the firing step
-            # (event_fired) and the propagation steps (add_event_propagation).
-            # Without it, different Event objects at the various method calls.
-	    attribute(:pending_events) { Hash.new }
-
-	    # Creates an Event object for +generator+, with the given argument
-	    # as parameters, or returns an already existing one
-	    def event_for(generator, event_id, time, context)
-		id, event = pending_events[generator]
-		if id && id == event_id
-		    return event
-		end
-		
-		event = generator.new(context)
-		event.send(:time=, time)
-		if generator.respond_to?(:task)
-		    generator.task.update_task_status(event)
-		end
-		pending_events[generator] = [event_id, event]
-		event
-	    end
-
-            # Message received when the +marshalled_from+ generator fired an
-            # event, with the given event id, time and context.
-	    def event_fired(marshalled_from, event_id, time, context)
-		from_generator = peer.local_object(marshalled_from)
-		context	       = peer.local_object(context)
-
-		event = event_for(from_generator, event_id, time, context)
-
-		event.send(:propagation_id=, from_generator.plan.engine.propagation_id)
-		from_generator.instance_variable_set("@happened", true)
-		from_generator.fired(event)
-		from_generator.call_handlers(event)
-
-		nil
-	    end
-
-            # Message received when the +marshalled_from+ generator has either
-            # been forwarded (only_forward = true) or signals (only_forward =
-            # false) the +marshalled_to+ generator. The remaining information
-            # describes the event itself.
-	    def event_add_propagation(only_forward, marshalled_from, marshalled_to, event_id, time, context)
-		from_generator  = peer.local_object(marshalled_from)
-		to_generator	= peer.local_object(marshalled_to)
-		context         = peer.local_object(context)
-
-		event		= event_for(from_generator, event_id, time, context)
-
-		# Only add the signalling if we own +to+
-		if to_generator.self_owned?
-                    to_generator.plan.engine.add_event_propagation(only_forward, [event], to_generator, event.context, nil)
-		else
-		    # Call #signalling or #forwarding to make
-		    # +from_generator+ look like as if the event was really
-		    # fired locally ...
-		    Distributed.update_all([from_generator.root_object, to_generator.root_object]) do
-			if only_forward then from_generator.forwarding(event, to_generator)
-			else from_generator.signalling(event, to_generator)
-			end
-		    end
-		end
-
-		nil
-	    end
-	end
-
         # This module defines the hooks required by dRoby on Roby::Task
 	module TaskNotifications
             # Hook called when the internal task data is modified. It sends
@@ -499,35 +318,5 @@ module Roby
 	    end
 	end
 	TaskArguments.include TaskArgumentsNotifications
-
-	class PeerServer
-            # Message received to announce that the internal data of +task+ is
-            # now +data+.
-	    def updated_data(task, data)
-		proxy = peer.local_object(task)
-		proxy.instance_variable_set("@data", peer.proxy(data))
-		nil
-	    end
-
-            # Message received to announce that the arguments of +task+ have
-            # been modified. +arguments+ is a hash containing only the new
-            # values.
-	    def updated_arguments(task, arguments)
-		proxy = peer.local_object(task)
-		arguments = peer.proxy(arguments)
-		Distributed.update(proxy) do
-		    proxy.arguments.merge!(arguments || {})
-		end
-		nil
-	    end
-	end
-
-	class PeerServer
-            # Message received to update our view of the remote robot state.
-	    def state_update(new_state)
-		peer.state = new_state
-		nil
-	    end
-	end
     end
 end
