+++ lib/roby/distributed/transaction.rb
@@ -433,62 +433,5 @@ module Roby
 		Roby::TaskEventGenerator::DRoby.new(controlable?, happened?, Distributed.format(task, dest), symbol)
 	    end
 	end
-
-	class PeerServer
-            # Message received when the 'prepare' stage of the transaction
-            # commit is requested.
-	    def transaction_prepare_commit(trsc)
-		trsc = peer.local_object(trsc)
-		peer.connection_space.transaction_prepare_commit(trsc)
-		trsc.freezed!
-		nil
-	    end
-            # Message received when a transaction commit is requested.
-	    def transaction_commit(trsc)
-		trsc = peer.local_object(trsc)
-		peer.connection_space.transaction_commit(trsc)
-		nil
-	    end
-            # Message received when a transaction commit is to be abandonned.
-	    def transaction_abandon_commit(trsc, error)
-		trsc = peer.local_object(trsc)
-		peer.connection_space.transaction_abandon_commit(trsc, error)
-		nil
-	    end
-            # Message received when a transaction discard is requested.
-	    def transaction_discard(trsc)
-		trsc = peer.local_object(trsc)
-		peer.connection_space.transaction_discard(trsc)
-		nil
-	    end
-            # Message received when the transaction edition token is given to
-            # this plan manager.
-	    def transaction_give_token(trsc, needs_edition)
-		trsc = peer.local_object(trsc)
-		trsc.edit!(needs_edition)
-		nil
-	    end
-	end
-
-	class Peer
-            # Send the information related to the given transaction in the
-            # remote plan manager.
-	    def transaction_propose(trsc)
-		synchro_point
-		create_sibling(trsc)
-		nil
-	    end
-
-            # Give the edition token on +trsc+ to the given peer.
-            # +needs_edition+ is a flag which, if true, requests that the token
-            # is given back at least once to the local plan manager.
-            #
-            # Do not use this directly, it is part of the multi-robot
-            # communication protocol. Use the edition-related methods on
-            # Distributed::Transaction instead.
-	    def transaction_give_token(trsc, needs_edition)
-		call(:transaction_give_token, trsc, needs_edition)
-	    end
-	end
     end
 end
